requirements.txt
google-genai
trafilatura
beautifulsoup4
requests